feat : "팀24 exception 기능 추가"
