+++ src/main/java/challenging/application/global/error/category/CategoryNotFoundException.java
@@ -1,21 +1,10 @@
 package challenging.application.global.error.category;
 
 import challenging.application.global.error.ErrorCode;
+import challenging.application.global.error.Team24Exception;
 
-public class CategoryNotFoundException extends RuntimeException {
-  private final ErrorCode errorCode;
-
+public class CategoryNotFoundException extends Team24Exception {
   public CategoryNotFoundException(ErrorCode errorCode) {
-    super(errorCode.getMessage());
-    this.errorCode = errorCode;
-  }
-
-  @Override
-  public String getMessage() {
-    return super.getMessage();
-  }
-
-  public int getStatusCode(){
-    return errorCode.getStatus().value();
+    super(errorCode);
   }
 }
