+++ src/main/java/challenging/application/global/error/challenge/AlreadyReservedException.java
@@ -1,21 +1,10 @@
 package challenging.application.global.error.challenge;
 
 import challenging.application.global.error.ErrorCode;
+import challenging.application.global.error.Team24Exception;
 
-public class AlreadyReservedException extends RuntimeException {
-  private final ErrorCode errorCode;
-
+public class AlreadyReservedException extends Team24Exception {
   public AlreadyReservedException(ErrorCode errorCode) {
-    super(errorCode.getMessage());
-    this.errorCode = errorCode;
-  }
-
-  @Override
-  public String getMessage() {
-    return super.getMessage();
-  }
-
-  public int getStatusCode(){
-    return errorCode.getStatus().value();
+    super(errorCode);
   }
 }
