+++ src/main/java/challenging/application/global/error/date/InvalidDateException.java
@@ -1,21 +1,10 @@
 package challenging.application.global.error.date;
 
 import challenging.application.global.error.ErrorCode;
+import challenging.application.global.error.Team24Exception;
 
-public class InvalidDateException extends RuntimeException {
-  private final ErrorCode errorCode;
-
+public class InvalidDateException extends Team24Exception {
   public InvalidDateException(ErrorCode errorCode) {
-    super(errorCode.getMessage());
-    this.errorCode = errorCode;
-  }
-
-  @Override
-  public String getMessage() {
-    return super.getMessage();
-  }
-
-  public int getStatusCode(){
-    return errorCode.getStatus().value();
+    super(errorCode);
   }
 }
