+++ src/main/java/challenging/application/global/error/history/HistoryNotFoundException.java
@@ -1,22 +1,10 @@
 package challenging.application.global.error.history;
 
 import challenging.application.global.error.ErrorCode;
+import challenging.application.global.error.Team24Exception;
 
-public class HistoryNotFoundException extends RuntimeException {
-  private final ErrorCode errorCode;
-
+public class HistoryNotFoundException extends Team24Exception {
   public HistoryNotFoundException(ErrorCode errorCode) {
-    super(errorCode.getMessage());
-    this.errorCode = errorCode;
+    super(errorCode);
   }
-
-  @Override
-  public String getMessage() {
-    return super.getMessage();
-  }
-
-  public int getStatusCode(){
-    return errorCode.getStatus().value();
-  }
-
 }
